pyproject.toml
[tool.poetry]
name = "fpgen"
version = "0.1.0"
description = ""
authors = ["phill.zotov <phill.zotov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.*"
click = "^8.0.1"
PyYAML = "^5.4.1"
Faker = "^8.16.0"

[tool.poetry.dev-dependencies]
SQLAlchemy = "^1.4.25"
SQLAlchemy-Utils = "^0.37.8"
psycopg2-binary = "^2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
